Feature/ci cd
Added ci cd for compiling the code pushed to dev and refactor(temporary) and main
also potential release support added
fixed lots of small warnings

+++ src/app/scenes/testScene.cpp
@@ -20,7 +20,7 @@ void TestScene::OnEnter(){
             row->Add(
                 Button(Text("Test", "TNR", 32, 0, RAYWHITE),
                     [j](){ dbg::GetLogger().Info("Button"+ std::to_string(j) +" clicked!"); },
-                    "button_default",{(float)200,(float)100}, 
+                    "button_default",{200.f,100.f}, 
                     UICSpec().SetFlex({.growth = 1, .shrink = 1})
                     .MinSize({100,50}))
             );
